perf: 홈 심사 목록을 SSR 병렬 페칭으로 전환하여 초기 로딩 워터폴 제거
홈 진입 시 LCP가 ~5초로 느린 문제를 개선한다.

기존에는 AuthProvider가 클라이언트에서 getUserProfile()을 호출(왕복1)해
스토어를 채운 뒤, 홈의 useQuery가 enabled: isAdmin !== undefined 조건에
막혀 있다가 프로필이 도착한 후에야 목록을 조회(왕복2)했다. 이 직렬
워터폴에 서버리스 콜드 스타트까지 겹쳐 첫 콘텐츠 렌더가 ~5초 지연됐다.
(LCP 4,955ms 중 99.7%가 render delay, 함수/Supabase 모두 서울 리전이라
지연 원인은 네트워크가 아닌 클라이언트 왕복 워터폴이었다.)

- page.tsx를 서버 컴포넌트로 전환(force-dynamic). 서버에서 프로필로
  role을 확정한 뒤 같은 리전의 Supabase로 심사 목록을 조회해 첫 화면
  데이터를 HTML에 포함시킨다. 브라우저 왕복 2~3회 → 1회로 단축.
- 검색/페이지네이션/필터 등 인터랙티브 로직은 HomeClient로 분리하고
  useQuery에 initialData를 시드한다. 초기 상태(1페이지·검색/필터 없음)는
  서버 데이터로 즉시 렌더하고, 이후 조작 시에만 클라이언트에서 재페치한다.

공용 액션(getAllJudgingWorkspaces) 시그니처와 AuthProvider, React Query
싱글톤은 변경하지 않아 다른 페이지에 영향이 없다.
